Game logic front/back integration

+++ backend/app/models/session.py
@@ -4,12 +4,13 @@ from typing import Optional, List
 from pydantic import BaseModel, Field
 
 
+
 class GameSession(BaseModel):
     """Game session model for Firestore."""
     user_id: str = Field(..., description="User's Firebase UID")
     session_id: str = Field(..., description="Unique session ID")
     route_id: str = Field(..., description="route ID")
-    total_earnings: float = Field(default=0.0, description="Total earnings in this session")
+    total_earnings: int = Field(default=0.0, description="Total earnings in this session")
     strikes: int = Field(default=0, description="Number of strikes (max 3)")
     questions_answered: int = Field(default=0, description="Number of questions answered")
     completed_at: Optional[datetime] = Field(default=None, description="Session completion timestamp")
@@ -18,3 +19,4 @@ class GameSession(BaseModel):
         json_encoders = {
             datetime: lambda v: v.isoformat()
         }
+
